refactor : 전체 패키지 구조 리펙터링
