+++ src/main/java/lotto/vo/CommonUnits.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.vo;
 
 public class CommonUnits {
     public static final int MONEY_UNIT = 1000;
