+++ src/main/java/lotto/vo/Grade.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.vo;
 
 public enum Grade {
